move method consume_wiki_summary to archive
