+++ README.md
@@ -1,5 +1,5 @@
 # Autocards
-* Automatically create questions and answers from various input formats (PDF files, webpages, wikipedia, epub files, etc) then export them into your favorite flashcards software (like [anki](https://apps.ankiweb.net/) or [SuperMemo](https://supermemo.guru/wiki/SuperMemo))
+* Automatically create questions and answers from various input formats (PDF files, webpages, epub files, etc) then export them into your favorite flashcards software (like [anki](https://apps.ankiweb.net/) or [SuperMemo](https://supermemo.guru/wiki/SuperMemo))
 * To see a real world example, the complete output of [this article](https://www.biography.com/political-figure/philip-ii-of-macedon) can be found [in this folder](./output_example/). It's the direct output with no post processing whatsoever.
 * Code is PEP compliant and with docstrings. Contributions and PR are extremely appreciated
 * Learn more by reading [the official write-up](https://psionica.org/docs/lab/autocards/).
@@ -20,7 +20,6 @@ a = Autocards()
 # eating the input text using one of the following ways:
 a.consume_var(my_text, per_paragraph=True) 
 a.consume_user_input(title="")
-a.consume_wiki_summary(keyword, lang="en")
 a.consume_textfile(filename, per_paragraph=True)
 a.consume_pdf(pdf_path, per_paragraph=True)
 a.consume_web(source, mode="url", element="p")
